third_party/WebKit/LayoutTests/inspector/tracing/timeline-paint-with-layout-invalidations-expected.txt

Tests the Timeline API instrumentation of paint events with layout invalidations.

PASS


Running: testLocalFrame
paint invalidations[
    {
        cause : {reason: Inline CSS style declaration was mutated, stackTrace: .../inspector/tracing/timeline-paint-with-layout-invalidations.html:10}
        changedAttribute : undefined
        changedClass : undefined
        changedId : undefined
        changedPseudo : undefined
        extraData : ""
        nodeName : "BODY"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
    {
        cause : {reason: Inline CSS style declaration was mutated, stackTrace: .../inspector/tracing/timeline-paint-with-layout-invalidations.html:11}
        changedAttribute : undefined
        changedClass : undefined
        changedId : undefined
        changedPseudo : undefined
        extraData : ""
        nodeName : "DIV id='testElement'"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
    {
        cause : {reason: Style changed, stackTrace: .../inspector/tracing/timeline-paint-with-layout-invalidations.html:12}
        changedAttribute : undefined
        changedClass : undefined
        changedId : undefined
        changedPseudo : undefined
        extraData : undefined
        nodeName : "DIV id='testElement'"
        selectorPart : undefined
        type : "LayoutInvalidationTracking"
    }
]

Running: testSubframe
second paint invalidations[
    {
        cause : {reason: Inline CSS style declaration was mutated, stackTrace: .../inspector/tracing/timeline-paint-with-layout-invalidations.html:21}
        changedAttribute : undefined
        changedClass : undefined
        changedId : undefined
        changedPseudo : undefined
        extraData : ""
        nodeName : "DIV"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
    {
        cause : {reason: Style changed, stackTrace: .../inspector/tracing/timeline-paint-with-layout-invalidations.html:22}
        changedAttribute : undefined
        changedClass : undefined
        changedId : undefined
        changedPseudo : undefined
        extraData : undefined
        nodeName : "DIV"
        selectorPart : undefined
        type : "LayoutInvalidationTracking"
    }
    {
        cause : {reason: Unknown, stackTrace: .../inspector/tracing/timeline-paint-with-layout-invalidations.html:22}
        changedAttribute : undefined
        changedClass : undefined
        changedId : undefined
        changedPseudo : undefined
        extraData : undefined
        nodeName : "#document"
        selectorPart : undefined
        type : "LayoutInvalidationTracking"
    }
]
